[MAINT] Rename and move testing utilities under tests directory in each module (#4047)
* Move dataset testing utils

* Fix path

* Rename decoding testing utils

* Rename glm testing utils

* Rename fmriprep testing utils

* Rename mass univariate testing utils

* Rename surface testing utils

* Formatting
