+++ nilearn/conftest.py
@@ -10,8 +10,8 @@ from nibabel import Nifti1Image
 from nilearn import image
 
 # we need to import these fixtures even if not used in this module
-from nilearn.datasets._testing import request_mocker  # noqa: F401
-from nilearn.datasets._testing import temp_nilearn_data_dir  # noqa: F401
+from nilearn.datasets.tests._testing import request_mocker  # noqa: F401
+from nilearn.datasets.tests._testing import temp_nilearn_data_dir  # noqa: F401
 
 collect_ignore = ["datasets/data/convert_templates.py"]
 collect_ignore_glob = ["reporting/_visual_testing/*"]
