+++ nilearn/datasets/tests/test_testing.py
@@ -8,7 +8,7 @@ import requests
 
 from nilearn import image
 from nilearn._utils.data_gen import generate_fake_fmri
-from nilearn.datasets import _testing
+from nilearn.datasets.tests import _testing
 
 
 def test_sender_key_order(request_mocker):
