+++ nilearn/glm/tests/_testing.py
@@ -1,3 +1,5 @@
+"""Test utilities for glm module."""
+
 import numpy as np
 import pandas as pd
