+++ nilearn/mass_univariate/tests/test_permuted_least_squares.py
@@ -55,7 +55,7 @@ def compare_to_ref_score(own_score, tested_var, target_var, covars=None):
 
 def ref_score(tested_var, target_var, covars=None):
     """Compute t-scores with linalg or statsmodels."""
-    from nilearn.mass_univariate.tests.utils import (
+    from nilearn.mass_univariate.tests._testing import (
         get_tvalue_with_alternative_library,
     )
