+++ pyproject.toml
@@ -121,7 +121,7 @@ extend-exclude = """
   | nilearn/interfaces/fmriprep/load_confounds_compcor.py
   | nilearn/interfaces/fmriprep/tests/test_load_confounds_utils.py
   | nilearn/interfaces/fmriprep/tests/test_load_confounds_strategy.py
-  | nilearn/interfaces/fmriprep/tests/utils.py
+  | nilearn/interfaces/fmriprep/tests/_testing.py
   | nilearn/interfaces/fmriprep/tests/test_load_confounds.py
   | nilearn/maskers/nifti_spheres_masker.py
   | nilearn/maskers/nifti_labels_masker.py
